Product Module - Manage Product in Back End (Part-2)

+++ app/Http/Controllers/Admin/AdminProductCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\ProductCategory;
+use App\Models\Product;
 
 class AdminProductCategoryController extends Controller
 {
@@ -42,10 +43,10 @@ class AdminProductCategoryController extends Controller
 
     public function destroy(Request $request, $id)
     {
-        // $check = Product::where('product_category_id', $id)->count();
-        // if($check > 0){
-        //     return back()->with('error', 'This product category is assigned to some products. You can not delete this product category.');
-        // }
+        $check = Product::where('product_category_id', $id)->count();
+        if($check > 0){
+            return back()->with('error', 'This product category is assigned to some products. You can not delete this product category.');
+        }
 
         $product_category = ProductCategory::where('id', $id)->first();
         $product_category->delete();
